feat: add maximum sync attempts
