+++ app/src/main/java/com/digiventure/ventnote/feature/note_backup/viewmodel/NoteBackupPageBaseVM.kt
@@ -1,5 +1,6 @@
-package com.digiventure.ventnote.feature.noteBackup.viewmodel
+package com.digiventure.ventnote.feature.note_backup.viewmodel
 
+import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.gms.auth.api.signin.GoogleSignInClient
@@ -11,13 +12,21 @@ interface NoteBackupPageBaseVM {
      * */
     val loader: MutableLiveData<Boolean>
 
-    val signInClient: GoogleSignInClient
+    val signInClient: GoogleSignInClient?
 
     val googleAccount: MutableLiveData<GoogleSignInAccount?>
+
+    val savedDay: LiveData<Long>
+
+    val maxAttempts: LiveData<Int>
 
     suspend fun backupDB(credential: GoogleAccountCredential): Result<Unit>
 
     suspend fun syncDB(credential: GoogleAccountCredential): Result<Unit>
+
+    suspend fun setSavedDay(previousDay: Long): Result<Unit>
+
+    suspend fun setMaxSyncAttempt(value: Int): Result<Unit>
 
     fun logout(): Result<Unit>
 }
