+++ app/src/main/java/com/digiventure/ventnote/feature/note_creation/components/AppBar.kt
@@ -1,9 +1,8 @@
-package com.digiventure.ventnote.feature.noteBackup.components
+package com.digiventure.ventnote.feature.note_creation.components
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material.icons.filled.Help
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -15,15 +14,14 @@ import com.digiventure.ventnote.components.navbar.TopNavBarIcon
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun NoteBackupAppBar(
+fun NoteCreationAppBar(
+    descriptionTextLength: Int,
     onBackPressed: () -> Unit,
-    onHelpPressed: () -> Unit,
     scrollBehavior: TopAppBarScrollBehavior) {
-
     TopAppBar(
         title = {
             Text(
-                text = "Backup Note",
+                text = if(descriptionTextLength > 0) "$descriptionTextLength" else "Add New Note",
                 color = MaterialTheme.colorScheme.onPrimary,
                 modifier = Modifier.padding(start = 8.dp),
             )
@@ -34,11 +32,6 @@ fun NoteBackupAppBar(
         navigationIcon = {
             TopNavBarIcon(Icons.Filled.ArrowBack, stringResource(R.string.back_nav_icon), Modifier.semantics {  }) {
                 onBackPressed()
-            }
-        },
-        actions = {
-            TopNavBarIcon(Icons.Default.Help, stringResource(R.string.menu_nav_icon), Modifier.semantics {  }) {
-                onHelpPressed()
             }
         },
         scrollBehavior = scrollBehavior,
