+++ app/src/main/java/com/digiventure/ventnote/feature/note_detail/NoteDetailPage.kt
@@ -1,37 +1,23 @@
-package com.digiventure.ventnote.feature.noteCreation
+package com.digiventure.ventnote.feature.note_detail
 
+import android.net.Uri
 import android.view.ViewTreeObserver
+import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ExtendedFloatingActionButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.OutlinedTextField
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.material3.TextFieldDefaults
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.material3.rememberTopAppBarState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.input.nestedscroll.nestedScroll
+import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.LocalView
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.contentDescription
@@ -47,58 +33,114 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.digiventure.ventnote.R
 import com.digiventure.ventnote.commons.TestTags
+import com.digiventure.ventnote.components.dialog.LoadingDialog
 import com.digiventure.ventnote.components.dialog.TextDialog
-import com.digiventure.ventnote.data.local.NoteModel
-import com.digiventure.ventnote.feature.noteCreation.components.NoteCreationAppBar
-import com.digiventure.ventnote.feature.noteCreation.viewmodel.NoteCreationPageBaseVM
-import com.digiventure.ventnote.feature.noteCreation.viewmodel.NoteCreationPageMockVM
-import com.digiventure.ventnote.feature.noteCreation.viewmodel.NoteCreationPageVM
+import com.digiventure.ventnote.feature.note_detail.components.NoteDetailAppBar
+import com.digiventure.ventnote.feature.note_detail.viewmodel.NoteDetailPageBaseVM
+import com.digiventure.ventnote.feature.note_detail.viewmodel.NoteDetailPageMockVM
+import com.digiventure.ventnote.feature.note_detail.viewmodel.NoteDetailPageVM
+import com.digiventure.ventnote.navigation.Route
 import com.digiventure.ventnote.ui.theme.PurpleGrey80
+import com.google.gson.Gson
 import kotlinx.coroutines.launch
 
-const val TAG : String = "NoteCreationPage"
+const val TAG : String = "NoteDetailPage"
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun NoteCreationPage(
+fun NoteDetailPage(
     navHostController: NavHostController,
-    viewModel: NoteCreationPageBaseVM = hiltViewModel<NoteCreationPageVM>()
+    viewModel: NoteDetailPageBaseVM = hiltViewModel<NoteDetailPageVM>(),
+    id: String
 ) {
     // String resource
-    val titleTextField = "${stringResource(R.string.title_textField)}-${TAG}"
-    val bodyTextField = "${stringResource(R.string.body_textField)}-${TAG}"
+    val titleTextField = "${stringResource(R.string.title_textField)}-$TAG"
+    val bodyTextField = "${stringResource(R.string.body_textField)}-$TAG"
     val titleInput = stringResource(R.string.title_textField_input)
     val bodyInput = stringResource(R.string.body_textField_input)
 
-    val length = viewModel.descriptionText.value.length
+    val noteDetailState = viewModel.noteDetail.observeAsState()
+    val data = noteDetailState.value?.getOrNull()
+
+    val isEditingState = viewModel.isEditing.value
+    val focusManager = LocalFocusManager.current
+
+    val loadingState = viewModel.loader.observeAsState()
 
     val scope = rememberCoroutineScope()
 
     val requiredDialogState = remember { mutableStateOf(false) }
+    val deleteDialogState = remember { mutableStateOf(false) }
     val cancelDialogState = remember { mutableStateOf(false) }
+    val openLoadingDialog = remember { mutableStateOf(false) }
     val snackBarHostState = remember { SnackbarHostState() }
+    
+    fun initData() {
+        viewModel.titleText.value = data?.title ?: ""
+        viewModel.descriptionText.value = data?.note ?: ""
+    }
 
-    fun addNote() {
-        if (viewModel.titleText.value.isEmpty() || viewModel.descriptionText.value.isEmpty()) {
-            requiredDialogState.value = true
-        } else {
+    LaunchedEffect(key1 = Unit) {
+        viewModel.getNoteDetail(id.toInt())
+    }
+
+    LaunchedEffect(key1 = noteDetailState.value) {
+        initData()
+    }
+
+    LaunchedEffect(key1 = isEditingState) {
+        if (!isEditingState) {
+            focusManager.clearFocus()
+        }
+    }
+
+    LaunchedEffect(key1 = loadingState.value) {
+        openLoadingDialog.value = loadingState.value == true
+    }
+
+    fun deleteNote() {
+        if (data != null) {
             scope.launch {
-                viewModel.addNote(NoteModel(
-                    title = viewModel.titleText.value,
-                    note = viewModel.descriptionText.value))
+                viewModel.deleteNoteList(data)
                     .onSuccess {
+                        deleteDialogState.value = false
                         navHostController.popBackStack()
-                        snackBarHostState.showSnackbar(
-                            message = "Note successfully added",
-                            withDismissAction = true
-                        )
                     }
                     .onFailure {
+                        deleteDialogState.value = false
                         snackBarHostState.showSnackbar(
                             message = it.message ?: "",
                             withDismissAction = true
                         )
                     }
+            }
+        }
+    }
+
+    fun updateNote() {
+        if (viewModel.titleText.value.isEmpty() || viewModel.descriptionText.value.isEmpty()) {
+            requiredDialogState.value = true
+        } else {
+            if (data != null) {
+                focusManager.clearFocus()
+
+                scope.launch {
+                    val updatedNote = data.copy(title = viewModel.titleText.value, note = viewModel.descriptionText.value)
+                    viewModel.updateNoteList(updatedNote)
+                        .onSuccess {
+                            viewModel.isEditing.value = false
+                            snackBarHostState.showSnackbar(
+                                message = "Note successfully updated",
+                                withDismissAction = true
+                            )
+                        }
+                        .onFailure {
+                            snackBarHostState.showSnackbar(
+                                message = it.message ?: "",
+                                withDismissAction = true
+                            )
+                        }
+                }
             }
         }
     }
@@ -132,28 +174,47 @@ fun NoteCreationPage(
 
     Scaffold(
         topBar = {
-            NoteCreationAppBar(
-                descriptionTextLength = length,
+            NoteDetailAppBar(
+                isEditing = isEditingState,
+                descriptionTextLength = viewModel.descriptionText.value.length,
                 onBackPressed = {
+                    navHostController.popBackStack()
+                },
+                onClosePressed = {
                     cancelDialogState.value = true
                 },
-                scrollBehavior = rememberedScrollBehavior
+                onDeletePressed = {
+                    deleteDialogState.value = true
+                },
+                scrollBehavior = rememberedScrollBehavior,
+                onSharePressed = {
+                    val json = Uri.encode(Gson().toJson(data))
+                    navHostController.navigate("${Route.SharePreviewPage.routeName}/${json}")
+                }
             )
         },
         snackbarHost = { SnackbarHost(snackBarHostState) },
         floatingActionButton = {
             ExtendedFloatingActionButton(
-                onClick = { addNote() },
+                onClick = {
+                    if(isEditingState) {
+                        updateNote()
+                    } else {
+                        viewModel.isEditing.value = true
+                    }
+                },
                 modifier = Modifier.semantics {
-                    testTag = "add-note-fab"
+                    testTag = "edit-note-fab"
                 },
                 text = {
-                    Text(stringResource(R.string.save), fontSize = 16.sp, fontWeight = FontWeight.Medium)
+                    Text(if(isEditingState) stringResource(R.string.save) else stringResource(R.string.edit),
+                        fontSize = 16.sp,
+                        fontWeight = FontWeight.Medium)
                 },
                 icon = {
                     Icon(
-                        imageVector = Icons.Filled.Check,
-                        contentDescription = stringResource(R.string.fab)
+                        imageVector = if(isEditingState) Icons.Filled.Check else Icons.Filled.Edit,
+                        contentDescription = stringResource(R.string.fab),
                     )
                 }
             )
@@ -174,6 +235,7 @@ fun NoteCreationPage(
                         },
                         textStyle = TextStyle(fontSize = 18.sp, fontWeight = FontWeight.Medium),
                         singleLine = true,
+                        readOnly = !isEditingState,
                         modifier = Modifier
                             .border(
                                 width = 3.dp,
@@ -191,6 +253,7 @@ fun NoteCreationPage(
                         },
                         textStyle = TextStyle(fontSize = 18.sp, fontWeight = FontWeight.Normal),
                         singleLine = false,
+                        readOnly = !isEditingState,
                         shape = RectangleShape,
                         colors = TextFieldDefaults.textFieldColors(
                             containerColor = Color.Transparent,
@@ -201,7 +264,7 @@ fun NoteCreationPage(
                         ),
                         modifier = Modifier
                             .fillMaxWidth()
-                            .fillMaxSize()
+                            .fillMaxHeight()
                             .semantics { contentDescription = bodyTextField },
                         placeholder = { Text(bodyInput, fontSize = 18.sp) }
                     )
@@ -209,7 +272,7 @@ fun NoteCreationPage(
             }
         },
         modifier = Modifier
-            .semantics { testTag = TestTags.NOTE_CREATION_PAGE }
+            .semantics { testTag = TestTags.NOTE_DETAIL_PAGE }
             .nestedScroll(scrollBehavior.nestedScrollConnection)
     )
 
@@ -234,16 +297,34 @@ fun NoteCreationPage(
         isOpened = cancelDialogState.value,
         onDismissCallback = { cancelDialogState.value = false },
         onConfirmCallback = {
-            navHostController.popBackStack()
+            viewModel.isEditing.value = false
             cancelDialogState.value = false
+
+            initData()
         })
+
+    TextDialog(
+        isOpened = deleteDialogState.value,
+        onDismissCallback = { deleteDialogState.value = false },
+        onConfirmCallback = { deleteNote() })
+
+    LoadingDialog(isOpened = openLoadingDialog.value, onDismissCallback = { openLoadingDialog.value = false })
+
+    BackHandler {
+        if (viewModel.isEditing.value) {
+            cancelDialogState.value = true
+        } else {
+            navHostController.popBackStack()
+        }
+    }
 }
 
 @Preview
 @Composable
-fun NoteCreationPagePreview() {
-    NoteCreationPage(
+fun NoteDetailPagePreview() {
+    NoteDetailPage(
         navHostController = rememberNavController(),
-        viewModel = NoteCreationPageMockVM()
+        viewModel = NoteDetailPageMockVM(),
+        id = "0"
     )
 }
