+++ app/src/main/java/com/digiventure/ventnote/module/DataStoreModule.kt
@@ -1,0 +1,45 @@
+package com.digiventure.ventnote.module
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.core.handlers.ReplaceFileCorruptionHandler
+import androidx.datastore.preferences.SharedPreferencesMigration
+import androidx.datastore.preferences.core.PreferenceDataStoreFactory
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.core.emptyPreferences
+import androidx.datastore.preferences.preferencesDataStoreFile
+import com.digiventure.ventnote.data.data_store.DataStoreHelper
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import javax.inject.Singleton
+
+private const val USER_PREFERENCES = "ventNote_preference"
+
+@Module
+@InstallIn(SingletonComponent::class)
+class DataStoreModule {
+    @Singleton
+    @Provides
+    fun dataStoreInstance(@ApplicationContext context: Context): DataStore<Preferences> {
+        return PreferenceDataStoreFactory.create(
+            corruptionHandler = ReplaceFileCorruptionHandler(
+                produceNewData = { emptyPreferences() }
+            ),
+            migrations = listOf(SharedPreferencesMigration(context, USER_PREFERENCES)),
+            scope = CoroutineScope(Dispatchers.IO + SupervisorJob()),
+            produceFile = { context.preferencesDataStoreFile(USER_PREFERENCES) }
+        )
+    }
+
+    @Singleton
+    @Provides
+    fun dataStoreHelper(dataStore: DataStore<Preferences>): DataStoreHelper {
+        return DataStoreHelper(dataStore)
+    }
+}
