+++ app/src/test/java/com/digiventure/ventnote/noteCreation/NoteCreationPageVMShould.kt
@@ -4,7 +4,7 @@ import com.digiventure.utils.BaseUnitTest
 import com.digiventure.utils.captureValues
 import com.digiventure.ventnote.data.NoteRepository
 import com.digiventure.ventnote.data.local.NoteModel
-import com.digiventure.ventnote.feature.noteCreation.viewmodel.NoteCreationPageVM
+import com.digiventure.ventnote.feature.note_creation.viewmodel.NoteCreationPageVM
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.test.runTest
