+++ app/src/test/java/com/digiventure/ventnote/noteDetail/NoteDetailPageVMShould.kt
@@ -5,7 +5,7 @@ import com.digiventure.utils.captureValues
 import com.digiventure.utils.getValueForTest
 import com.digiventure.ventnote.data.NoteRepository
 import com.digiventure.ventnote.data.local.NoteModel
-import com.digiventure.ventnote.feature.noteDetail.viewmodel.NoteDetailPageVM
+import com.digiventure.ventnote.feature.note_detail.viewmodel.NoteDetailPageVM
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.runBlocking
